projet_migrations/streaming/processor/requirements.txt
confluent-kafka==2.6.0
requests==2.32.3
python-dateutil==2.9.0.post0
numpy==1.26.4
pandas==2.2.2
